requirements.txt
redis>=5.0.0
orjson>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.66.0
openai>=1.41.0
typing_extensions>=4.8.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
tensorflow[and-cuda]>=2.13.0  
transformers>=4.30.0
tf-keras